requirement.txt
customtkinter
requests
pygame
groq
streamlit
python-dotenv